requirements.txt
flask
tensorflow
numpy
Pillow
gunicorn
uvicorn